src/main/java/lol/koblizek/notecz/api/user/UserService.java
package lol.koblizek.notecz.api.user;

import jakarta.validation.ConstraintViolationException;
import jakarta.validation.Valid;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Service;
import org.springframework.validation.annotation.Validated;
import org.springframework.validation.method.MethodValidationException;

import java.util.Optional;

@Service
@Validated
public class UserService {

    private static final Logger LOGGER = LoggerFactory.getLogger(UserService.class);

    private final UserRepository userRepository;

    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public Optional<User> getUserById(Long id) {
        return userRepository.findById(id);
    }

    public Optional<User> save(@Valid User user) {
        return Optional.ofNullable(userRepository.saveAndFlush(user));
    }
}
